pyproject.toml
[project]
name = "fastapi-ca"
version = "0.1.0"
description = ""
authors = [
    {name = "pinstinct",email = "contact@limhm.dev"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "py-ulid (>=1.0.3,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "dependency-injector (>=4.46.0,<5.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "bcrypt (==3.2.0)",
    "celery (>=5.5.2,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "freezegun (>=1.5.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.3.2"
isort = "^6.0.1"
pre-commit = "^4.2.0"


[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
skip = [".venv", "build", "dist"]

[tool.autopep8]
max_line_length = 88
aggressive = 2
